limpando o app.tsx e passando a funçao do provider para o proprio PlayerContext.tsx
