+++ src/pages/_app.tsx
@@ -2,32 +2,14 @@ import '../styles/global.scss'
 
 import { Header } from '../components/Header'
 import { Player } from '../components/Player';
-import { PlayerContext } from '../contexts/PlayerContext';
 
 import styles from '../styles/app.module.scss';
 import React, { useState } from 'react';
+import { PlayerContextProvider } from '../contexts/PlayerContext';
 
 function MyApp({ Component, pageProps }) {
-  const [ episodeList, setEpisodeList ] = useState([]);
-  const [ currentEpisodeIndex, setCurrentEpisodeIndex ] = useState(0);
-  const [ isPlaying, setIsPlaying ] = useState(false);
-
-  function play(episode) {
-    setEpisodeList([episode]);
-    setCurrentEpisodeIndex(0);
-    setIsPlaying(true);
-  }
-
-  function togglePlay() {
-    setIsPlaying(!isPlaying);
-  }
-
-  function setPlayingState(state: boolean) {
-    setIsPlaying(state);
-  }
-
   return (
-    <PlayerContext.Provider value={{ episodeList, currentEpisodeIndex, play, isPlaying, togglePlay, setPlayingState }}>
+    <PlayerContextProvider>
       <div className={styles.wrapper}>
         <main>
           <Header />
@@ -35,7 +17,7 @@ function MyApp({ Component, pageProps }) {
         </main>
         <Player />
       </div>
-    </PlayerContext.Provider>
+    </PlayerContextProvider>
   )
 }
